refactor: create constants module and shuffle imports
