+++ linkup-cli/src/constants.rs
@@ -1,0 +1,8 @@
+pub const LINKUP_CONFIG_ENV: &str = "LINKUP_CONFIG";
+pub const LINKUP_LOCALSERVER_PORT: u16 = 9066;
+pub const LINKUP_DIR: &str = ".linkup";
+pub const LINKUP_STATE_FILE: &str = "state";
+pub const LINKUP_LOCALSERVER_PID_FILE: &str = "localserver-pid";
+pub const LINKUP_CLOUDFLARED_PID: &str = "cloudflared-pid";
+pub const LINKUP_LOCALDNS_INSTALL: &str = "localdns-install";
+pub const LINKUP_CF_TLS_API_ENV_VAR: &str = "LINKUP_CF_API_TOKEN";
