+++ linkup-cli/src/local_config.rs
@@ -1,8 +1,5 @@
-use std::{
-    env,
-    fmt::{self, Display, Formatter},
-    fs,
-};
+use std::fmt::{self, Display, Formatter};
+use std::{env, fs};
 
 use rand::{distributions::Alphanumeric, Rng};
 use serde::{Deserialize, Serialize};
@@ -10,7 +7,8 @@ use url::Url;
 
 use linkup::{CreatePreviewRequest, StorableDomain, StorableRewrite, StorableService};
 
-use crate::{linkup_file_path, CliError, LINKUP_CONFIG_ENV, LINKUP_STATE_FILE};
+use crate::constants::{LINKUP_CONFIG_ENV, LINKUP_STATE_FILE};
+use crate::{linkup_file_path, CliError};
 
 #[derive(Deserialize, Serialize, Clone)]
 pub struct LocalState {
